[chore] Update to skip the gitignore when the xml files are cleaned

+++ tornado-assembly/src/bin/idea_xml_utils.py
@@ -201,6 +201,8 @@ def cleanup_build_directory(buildDirectoryString):
     tornadoRootDirectory = Path(buildDirectoryString)
     if tornadoRootDirectory.exists() and tornadoRootDirectory.is_dir():
         for file in tornadoRootDirectory.iterdir():
+            if file.name == ".gitignore":
+                continue
             if file.is_file():
                 file.unlink()
